move of BaseCrawlingStrategy, BFS/DFS strategy
