+++ frontera/strategy/__init__.py
@@ -1,9 +1,5 @@
-# -*- coding: utf-8 -*-
-from __future__ import absolute_import
-from frontera.core.models import Request
-from frontera.contrib.middlewares.fingerprint import UrlFingerprintMiddleware
-
 from abc import ABCMeta, abstractmethod
+
 import six
