+++ frontera/strategy/depth.py
@@ -1,11 +1,10 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
-from six.moves.urllib.parse import urlparse
 from frontera.core.components import States
-from frontera.worker.strategies import BaseCrawlingStrategy
+from frontera.strategy import BaseCrawlingStrategy
 
 
-class CrawlingStrategy(BaseCrawlingStrategy):
+class BreadthFirstCrawlingStrategy(BaseCrawlingStrategy):
     def read_seeds(self, fh):
         for url in fh:
             url = url.strip()
@@ -13,6 +12,7 @@ class CrawlingStrategy(BaseCrawlingStrategy):
             self.refresh_states(req)
             if req.meta[b'state'] is States.NOT_CRAWLED:
                 req.meta[b'state'] = States.QUEUED
+                req.meta[b'depth'] = 0
                 self.schedule(req)
 
     def page_crawled(self, response):
@@ -23,15 +23,21 @@ class CrawlingStrategy(BaseCrawlingStrategy):
 
     def links_extracted(self, request, links):
         for link in links:
+            link.meta[b'depth'] = request.meta[b'depth'] + 1
             if link.meta[b'state'] is States.NOT_CRAWLED:
                 link.meta[b'state'] = States.QUEUED
-                self.schedule(link, self.get_score(link.url))
+                self.schedule(link, self.get_score(link))
 
     def page_error(self, request, error):
         request.meta[b'state'] = States.ERROR
         self.schedule(request, score=0.0, dont_queue=True)
 
-    def get_score(self, url):
-        url_parts = urlparse(url)
-        path_parts = url_parts.path.split('/')
-        return 1.0 / (max(len(path_parts), 1.0) + len(url_parts.path) * 0.1)
+    def get_score(self, link):
+        depth = float(link.meta[b'depth'])
+        return 1.0 - (depth / (depth + 1.0))
+
+
+class DepthFirstCrawlingStrategy(BreadthFirstCrawlingStrategy):
+    def get_score(self, link):
+        depth = float(link.meta[b'depth'])
+        return depth / (depth + 1.0)
